requirements.txt
fastapi
uvicorn
requests
python-dateutil
icalendar
recurring-ical-events
